+++ doc/source/history.rst
@@ -26,6 +26,8 @@ Release history
 
 * CHANGED: signature of :py:meth:`~reikna.core.operation.OperationRecorder.add_kernel` takes a renderable instead of a full template.
 
+* CHANGED: :py:meth:`~reikna.cluda.api.Context.compile_static` now takes a template instead of a source.
+
 * FIXED: some OpenCL compilation warnings caused by an incorrect version querying macro.
 
 * FIXED: bug with incorrect processing of scalar global size in static kernels.